+++ CHANGELOG.md
@@ -6,12 +6,12 @@ file.
 ## [Unreleased]
 ### Added
 - Working llvm coverage instrumentation (coverage now works on Mac and Windows!)
-- `--post-test-delay` for tests which spawn a process
+- `--post-test-delay` for tests which spawn a process (default 1s for llvm coverage)
 
 ### Changed
 - Update quick-xml to 0.25
 - Make --ignore-tests the default and add a flag --include-tests to reapply the old behaviour
-- profraw files moved to `$TARGET_DIR/tarpaulin/profraws`
+- profraw files moved to `$TARGET_DIR/tarpaulin/profraws` (configurable via `Config::set_profraw_folder`)
 
 ## [0.21.0] 2022-08-30
 ### Changed